pyproject.toml
[tool.poetry]
name = "event-horizon-api"
version = "0.1.0"
description = ""
authors = ["Nicolas Epstein <nicolas.epstein@qc.croixbleue.ca>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.75.2"}
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
pytest = "^7.1.2"
requests = "^2.27.1"
python-multipart = "^0.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
